done collecting parameter values and sending them over to the agaveApp
* The RUN button in the parameter tab will collect all the current
parameter values, put them into a QMap<QString, QString> object, and
emit a signal that should be handles by the agave App.
* I implemented an internal variable to hold that information inside
the CFDagaveApps, as well as the public slot to receive the parameter
information.
* PETER S., please look at where to best place the connect for this
action.  I put a (commented out) template for the connect statement
into the constructor for the CFDagaveApps, but it does not belong there.